refactor(audit): migrate MinIO and Database to unified AuditEvent model
- Remove deprecated MinIOAuditLog and DatabaseAuditLog models
- Both subsystems now use unified audit.AsyncLogger[*models.AuditEvent]
- Delete 9 legacy files (models, repos, services, tests, handlers)
- Create sync wrapper for backward compatibility
- Update scheduler task to use unified repository
- Audit data now segregated by subsystem field (minio/database)

Completed tasks: T036, T037
