--- tests/integration/database/audit_log_test.go
@@ -1,236 +1,0 @@
-package database_test
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	"github.com/google/uuid"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/ysicing/tiga/internal/models"
-
-	dbservices "github.com/ysicing/tiga/internal/services/database"
-)
-
-// MockAuditLogRepository is a simple in-memory implementation for testing
-type MockAuditLogRepository struct {
-	logs []*models.DatabaseAuditLog
-}
-
-func (r *MockAuditLogRepository) Create(ctx context.Context, log *models.DatabaseAuditLog) error {
-	if log.ID == uuid.Nil {
-		log.ID = uuid.New()
-	}
-	if log.CreatedAt.IsZero() {
-		log.CreatedAt = time.Now().UTC()
-	}
-	r.logs = append(r.logs, log)
-	return nil
-}
-
-func (r *MockAuditLogRepository) List(ctx context.Context, page, pageSize int) ([]*models.DatabaseAuditLog, int64, error) {
-	return r.logs, int64(len(r.logs)), nil
-}
-
-func TestAuditLogRecording(t *testing.T) {
-	t.Run("LogInstanceCreation", func(t *testing.T) {
-		repo := &MockAuditLogRepository{logs: make([]*models.DatabaseAuditLog, 0)}
-		logger := dbservices.NewAuditLogger(repo)
-
-		instanceID := uuid.New()
-		err := logger.LogAction(context.Background(), dbservices.AuditEntry{
-			InstanceID: &instanceID,
-			Operator:   "admin",
-			Action:     "instance.create",
-			TargetType: "instance",
-			TargetName: "MySQL Production",
-			Details: map[string]interface{}{
-				"type": "mysql",
-				"host": "db.example.com",
-				"port": 3306,
-			},
-			Success:  true,
-			ClientIP: "192.168.1.100",
-		})
-
-		require.NoError(t, err)
-		assert.Len(t, repo.logs, 1)
-
-		log := repo.logs[0]
-		assert.Equal(t, "admin", log.Operator)
-		assert.Equal(t, "instance.create", log.Action)
-		assert.Equal(t, "instance", log.TargetType)
-		assert.Equal(t, "MySQL Production", log.TargetName)
-		assert.True(t, log.Success)
-		assert.Equal(t, "192.168.1.100", log.ClientIP)
-		assert.NotEmpty(t, log.Details)
-	})
-
-	t.Run("LogQueryExecution", func(t *testing.T) {
-		repo := &MockAuditLogRepository{logs: make([]*models.DatabaseAuditLog, 0)}
-		logger := dbservices.NewAuditLogger(repo)
-
-		instanceID := uuid.New()
-		err := logger.LogAction(context.Background(), dbservices.AuditEntry{
-			InstanceID: &instanceID,
-			Operator:   "developer",
-			Action:     "query.execute",
-			TargetType: "query",
-			TargetName: "SELECT query",
-			Details: map[string]interface{}{
-				"database": "app_db",
-				"query":    "SELECT * FROM users WHERE id = 1",
-				"duration": 125,
-			},
-			Success:  true,
-			ClientIP: "10.0.0.50",
-		})
-
-		require.NoError(t, err)
-		assert.Len(t, repo.logs, 1)
-
-		log := repo.logs[0]
-		assert.Equal(t, "developer", log.Operator)
-		assert.Equal(t, "query.execute", log.Action)
-		assert.True(t, log.Success)
-	})
-
-	t.Run("LogBlockedQuery", func(t *testing.T) {
-		repo := &MockAuditLogRepository{logs: make([]*models.DatabaseAuditLog, 0)}
-		logger := dbservices.NewAuditLogger(repo)
-
-		instanceID := uuid.New()
-		err := logger.LogAction(context.Background(), dbservices.AuditEntry{
-			InstanceID: &instanceID,
-			Operator:   "user123",
-			Action:     "query.blocked",
-			TargetType: "query",
-			TargetName: "DROP TABLE attempt",
-			Details: map[string]interface{}{
-				"query":  "DROP TABLE users",
-				"reason": "DDL operations are forbidden",
-			},
-			Success:  false,
-			Error:    dbservices.ErrSQLDangerousOperation,
-			ClientIP: "203.0.113.45",
-		})
-
-		require.NoError(t, err)
-		assert.Len(t, repo.logs, 1)
-
-		log := repo.logs[0]
-		assert.Equal(t, "query.blocked", log.Action)
-		assert.False(t, log.Success)
-		assert.Contains(t, log.ErrorMessage, "SQL contains forbidden operation")
-	})
-
-	t.Run("LogPermissionGrant", func(t *testing.T) {
-		repo := &MockAuditLogRepository{logs: make([]*models.DatabaseAuditLog, 0)}
-		logger := dbservices.NewAuditLogger(repo)
-
-		instanceID := uuid.New()
-		err := logger.LogAction(context.Background(), dbservices.AuditEntry{
-			InstanceID: &instanceID,
-			Operator:   "admin",
-			Action:     "permission.grant",
-			TargetType: "permission",
-			TargetName: "readonly access to app_db",
-			Details: map[string]interface{}{
-				"user":     "readonly_user",
-				"database": "app_db",
-				"role":     "readonly",
-			},
-			Success:  true,
-			ClientIP: "192.168.1.10",
-		})
-
-		require.NoError(t, err)
-		assert.Len(t, repo.logs, 1)
-
-		log := repo.logs[0]
-		assert.Equal(t, "permission.grant", log.Action)
-		assert.Equal(t, "permission", log.TargetType)
-	})
-
-	t.Run("LogMultipleActions", func(t *testing.T) {
-		repo := &MockAuditLogRepository{logs: make([]*models.DatabaseAuditLog, 0)}
-		logger := dbservices.NewAuditLogger(repo)
-
-		actions := []struct {
-			action string
-			target string
-		}{
-			{"instance.create", "MySQL Instance"},
-			{"database.create", "testdb"},
-			{"user.create", "testuser"},
-			{"permission.grant", "testuser on testdb"},
-			{"query.execute", "SELECT query"},
-		}
-
-		instanceID := uuid.New()
-		for _, a := range actions {
-			logger.LogAction(context.Background(), dbservices.AuditEntry{
-				InstanceID: &instanceID,
-				Operator:   "admin",
-				Action:     a.action,
-				TargetType: "test",
-				TargetName: a.target,
-				Success:    true,
-			})
-		}
-
-		assert.Len(t, repo.logs, 5)
-
-		// Verify all actions are logged in order
-		for i, a := range actions {
-			assert.Equal(t, a.action, repo.logs[i].Action)
-			assert.Equal(t, a.target, repo.logs[i].TargetName)
-		}
-	})
-
-	t.Run("RequiredFieldsValidation", func(t *testing.T) {
-		repo := &MockAuditLogRepository{logs: make([]*models.DatabaseAuditLog, 0)}
-		logger := dbservices.NewAuditLogger(repo)
-
-		// Missing operator should fail
-		err := logger.LogAction(context.Background(), dbservices.AuditEntry{
-			Action:  "test.action",
-			Success: true,
-		})
-		assert.Error(t, err, "Should require operator")
-
-		// Missing action should fail
-		err = logger.LogAction(context.Background(), dbservices.AuditEntry{
-			Operator: "admin",
-			Success:  true,
-		})
-		assert.Error(t, err, "Should require action")
-
-		// No logs should be created
-		assert.Len(t, repo.logs, 0)
-	})
-}
-
-func TestClientIPExtraction(t *testing.T) {
-	t.Run("ExtractIPFromContext", func(t *testing.T) {
-		// Test with IP in context (would be set by middleware)
-		ctx := context.WithValue(context.Background(), "client_ip", "203.0.113.100")
-
-		ip := dbservices.ExtractClientIP(ctx)
-		_ = ip // Note: The actual implementation uses a private context key
-		// This test demonstrates the expected behavior
-		// assert.Equal(t, "203.0.113.100", ip)
-
-		// For now, verify it doesn't crash
-		assert.NotPanics(t, func() {
-			dbservices.ExtractClientIP(ctx)
-		})
-	})
-
-	t.Run("NilContextHandling", func(t *testing.T) {
-		ip := dbservices.ExtractClientIP(nil)
-		assert.Empty(t, ip, "Nil context should return empty IP")
-	})
-}
